Skip async actions if table does not exist
Two async actions `purge_cache` and `check_mem_usage` are scheduled
using `timer:send_after` and `timer:apply_after` interfaces, which
run whether or not the cache ETS tables are still there, causing
spurious crashes when caches are stopped or abandoned. This adds
checks to simply skip the async actions when the table is missing.