disable DEBUG logs, log purge_cache at DEBUG level
because #8 moved erl_cache from `lager` to `error_logger`, debug level logging is no longer possible. Instead of promoting it to info level, this change simply squelches it for now.

The cache purge message is annoying in caches that are short-lived. There is no reason to log this data at the info level. Log it at the debug level, so that, if debug logs are enabled in the future, it will be logged appropriately.